+++ src/GenCondField.js
@@ -3,12 +3,12 @@ import React, {Component} from 'react';
 import omit from 'lodash/omit';
 
 import {consumeGenContext} from './contextUtils';
-import {evalCond, evalCondRequired} from './conditionalUtils';
+import {evalCond} from './conditionalUtils';
 import GenField from './GenField';
 
 class GenCondField extends Component {
   render() {
-    const {field, _parentValue, parentVisible, path, gen: {customFieldTypes}, ...fields} = this.props;
+    const {field, parentQuestionId, parentVisible, path, gen: {customFieldTypes}, ...data} = this.props;
     return (
       <GenField
         {...{
@@ -19,34 +19,28 @@ class GenCondField extends Component {
             (field.conditionalVisible
               ? evalCond({
                   cond: field.conditionalVisible,
-                  data: {
-                    ...fields,
-                    _parentValue
-                  },
+                  data,
                   customFieldTypes,
-                  reduxFormDeep: true
+                  reduxFormDeep: true,
+                  ...(parentQuestionId && {valueKey: parentQuestionId})
                 })
               : true),
           required: field.conditionalRequired
-            ? evalCondRequired({
+            ? evalCond({
                 cond: field.conditionalRequired,
-                data: {
-                  ...fields,
-                  _parentValue
-                },
+                data,
                 customFieldTypes,
-                reduxFormDeep: true
+                reduxFormDeep: true,
+                ...(parentQuestionId && {valueKey: parentQuestionId})
               })
             : false,
           disabled: field.conditionalDisabled
             ? evalCond({
                 cond: field.conditionalDisabled,
-                data: {
-                  ...fields,
-                  _parentValue
-                },
+                data,
                 customFieldTypes,
-                reduxFormDeep: true
+                reduxFormDeep: true,
+                ...(parentQuestionId && {valueKey: parentQuestionId})
               })
             : false
         }}